OK-509: Refaktoroitu ListTable-sarakkeiden muodostamiseen liittyvä koodi omaan moduuliinsa
