+++ src/app/components/table/list-table.tsx
@@ -10,129 +10,13 @@ import {
   styled,
 } from '@mui/material';
 import { useTranslations } from '@/app/hooks/useTranslations';
-import { TFunction } from 'i18next';
-import { ExternalLink } from '../external-link';
 import { useMemo } from 'react';
 import { OphPagination } from './oph-pagination';
 import { ophColors } from '@opetushallitus/oph-design-system';
 import { TableHeaderCell } from './table-header-cell';
 import { EMPTY_ARRAY, EMPTY_STRING_SET } from '@/app/lib/common';
 import { TableHeaderCheckbox, TableRowCheckbox } from './table-checkboxes';
-import { KeysMatching, ListTableColumn, Row } from './table-types';
-
-export const makeGenericColumn = <T extends Record<string, unknown>>({
-  title,
-  key,
-  valueProp,
-}: {
-  title: string;
-  key: string;
-  valueProp: KeysMatching<T, string | number | undefined>;
-}): ListTableColumn<T> => ({
-  title,
-  key,
-  render: (props) => <span>{props[valueProp] as string}</span>,
-  style: { width: 'auto' },
-});
-
-export const makeColumnWithCustomRender = <T extends Record<string, unknown>>({
-  title,
-  key,
-  renderFn,
-  sortable = true,
-}: {
-  title: string;
-  key: string;
-  renderFn: (props: T) => React.ReactNode;
-  sortable?: boolean;
-}): ListTableColumn<T> => ({
-  title,
-  key,
-  render: (props) => renderFn(props),
-  sortable,
-});
-
-export const makeBooleanYesNoColumn = <T extends Record<string, unknown>>({
-  t,
-  title,
-  key,
-  booleanValueProp,
-}: {
-  t: TFunction;
-  title: string;
-  key: string;
-  booleanValueProp: KeysMatching<T, boolean>;
-}): ListTableColumn<T> => ({
-  title,
-  key,
-  render: (props) => (
-    <span>
-      {(props[booleanValueProp] as boolean)
-        ? t('yleinen.kylla')
-        : t('yleinen.ei')}
-    </span>
-  ),
-  style: { width: 'auto' },
-});
-
-export const makeColumnWithValueToTranslate = <
-  T extends Record<string, unknown>,
->({
-  t,
-  title,
-  key,
-  valueProp,
-}: {
-  t: TFunction;
-  title: string;
-  key: string;
-  valueProp: KeysMatching<T, string>;
-}): ListTableColumn<T> => ({
-  title,
-  key,
-  render: (props) => <span>{t(props[valueProp] as string)}</span>,
-  style: { width: 'auto' },
-});
-
-export const makeCountColumn = <T extends Record<string, unknown>>({
-  title,
-  key,
-  amountProp,
-}: {
-  title: string;
-  key: string;
-  amountProp: KeysMatching<T, number>;
-}): ListTableColumn<T> => ({
-  title,
-  key,
-  render: (props) => <span>{(props[amountProp] ?? 0) as number}</span>,
-  style: { width: 0 },
-});
-
-export const makeExternalLinkColumn = <T extends Record<string, unknown>>({
-  linkBuilder,
-  title,
-  key,
-  nameProp,
-  linkProp,
-}: {
-  linkBuilder: (s: string) => string;
-  title: string;
-  key: string;
-  nameProp?: KeysMatching<T, string>;
-  linkProp: KeysMatching<T, string>;
-}): ListTableColumn<T> => ({
-  title,
-  key,
-  render: (props) => (
-    <ExternalLink
-      noIcon={true}
-      name={props[nameProp ?? linkProp] as string}
-      href={linkBuilder(props[linkProp] as string)}
-    />
-  ),
-  style: { width: 'auto' },
-});
+import { ListTableColumn, Row } from './table-types';
 
 const StyledTable = styled(Table)({
   width: '100%',
@@ -323,5 +207,3 @@ export const ListTable = <T extends Row>({
     </TableWrapper>
   );
 };
-
-export default ListTable;
