+++ src/app/haku/[oid]/hakukohde/[hakukohde]/perustiedot/valintatapajonot-table.tsx
@@ -1,9 +1,10 @@
 'use client';
 
-import ListTable, {
+import { ListTable } from '@/app/components/table/list-table';
+import {
   makeCountColumn,
   makeGenericColumn,
-} from '@/app/components/table/list-table';
+} from '@/app/components/table/table-columns';
 import { isToisenAsteenYhteisHaku } from '@/app/lib/kouta';
 import { Haku } from '@/app/lib/types/kouta-types';
 import { ValintatapajonoTulos } from '@/app/lib/types/sijoittelu-types';
