+++ src/app/haku/[oid]/hakukohde/[hakukohde]/valintakoekutsut/valintakoekutsut-table.tsx
@@ -1,15 +1,16 @@
 'use client';
 import { ValintakoekutsutActionBar } from './valintakoekutsut-action-bar';
-import ListTable, {
-  makeColumnWithValueToTranslate,
-} from '@/app/components/table/list-table';
 import { useTranslations } from '@/app/hooks/useTranslations';
 import { Box } from '@mui/material';
 import { useMemo, useState } from 'react';
 import { toFormattedDateTimeString } from '@/app/lib/localization/translation-utils';
 import { ValintakoeKutsuItem } from '@/app/lib/types/valintakoekutsut-types';
-import { hakijaColumn } from '@/app/components/table/hakija-column';
 import { ListTableColumn } from '@/app/components/table/table-types';
+import {
+  hakijaColumn,
+  makeColumnWithValueToTranslate,
+} from '@/app/components/table/table-columns';
+import { ListTable } from '@/app/components/table/list-table';
 
 const TRANSLATIONS_PREFIX = 'valintakoekutsut.taulukko';
